+++ client/components/Button/index.js
@@ -1,56 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
-import styled from 'styled-components'
 import { withHandlers } from 'recompose'
 
-const Wrapper = styled.button`
-  border: 10px;
-  box-sizing: border-box;
-  display: inline-block;
-  font-family: Roboto, sans-serif;
-  -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-  cursor: pointer;
-  text-decoration: none;
-  margin: 0;
-  padding: 0;
-  outline: none;
-  font-size: inherit;
-  font-weight: 400;
-  position: relative;
-  z-index: 1;
-  height: 36px;
-  line-height: 36px;
-  border-radius: 2px;
-  transition: all ${props => props.theme.animation.default};
-  text-align: center;
-  color: ${props => props.theme.color.text};
-
-  &.fullWidth {
-    width: 100%;
-  }
-
-  &.primary {
-    background: ${props => props.theme.color.primary};
-    color: ${props => props.theme.color.textAlternate};
-  }
-
-  &.secondary {
-    background: ${props => props.theme.color.secondary};
-    color: ${props => props.theme.color.textAlternate};
-  }
-`
-
-const InnerOverlay = styled.div`
-  width: 100%;
-  height: 100%;
-  padding: 0 24px;
-  transition: all ${props => props.theme.animation.default};
-
-  &:hover {
-    background-color: rgba(255, 255, 255, 0.2);
-  }
-`
+import Wrapper from './Wrapper'
+import InnerOverlay from './InnerOverlay'
 
 const Button = ({
   className,
