+++ client/components/Loading/fadeIn.js
@@ -1,0 +1,8 @@
+import { keyframes } from 'styled-components'
+
+const fadeIn = keyframes`
+  from { opacity: 0 }
+  to { opacity: 0.7 }
+`
+
+export default fadeIn
