+++ client/components/Loading/rotate.js
@@ -1,0 +1,8 @@
+import { keyframes } from 'styled-components'
+
+const rotate = keyframes`
+  from { transform: rotate(45deg) }
+  to { transform: rotate(405deg) }
+`
+
+export default rotate
