+++ client/components/NavigationIcon/Icon.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const Icon = styled.div`
+  position: relative;
+  transform: rotate(0deg);
+  transition: ${props => props.theme.animation.default};
+`
+
+export default Icon
