+++ client/components/NavigationItemList/Wrapper.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+  width: 100%;
+`
+
+export default Wrapper
