+++ client/components/NavigationMenu/Divider.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Divider = styled.div`
+  height: 1px;
+  width: 100%;
+  margin: 8px 0;
+  background: ${props => props.theme.color.grey4};
+`
+
+export default Divider
