+++ client/components/PostList/Wrapper.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  padding-bottom: 10px;
+`
+
+export default Wrapper
