+++ client/pages/Home/Home.js
@@ -1,12 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styled from 'styled-components'
 import { createFragmentContainer, graphql } from 'react-relay'
 
-const Wrapper = styled.div`
-  margin-top: 20px;
-  text-align: center;
-`
+import Wrapper from './Wrapper'
 
 const HomePage = ({ viewer }) => (
   <Wrapper>
