+++ client/pages/Home/Wrapper.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+  margin-top: 20px;
+  text-align: center;
+`
+export default Wrapper
