+++ client/pages/UserCreatePost/Form.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+import Formsy from 'formsy-react'
+
+const Form = styled(Formsy.Form)`
+  max-width: 600px;
+  width: 90%;
+  margin-left: auto;
+  margin-right: auto;
+`
+
+export default Form
