+++ client/pages/UserLogin/Bold.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Bold = styled.b`
+  font-weight: 400;
+`
+
+export default Bold
