+++ client/pages/UserLogin/Hint.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Hint = styled.div`
+  max-width: 400px;
+  margin: auto;
+  font-size: 14px;
+  line-height: 20px;
+`
+
+export default Hint
