+++ client/pages/UserProfile/Wrapper.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+  padding: 10px;
+`
+
+export default Wrapper
