+++ client/pages/UserRegister/Form.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+import Formsy from 'formsy-react'
+
+const Form = styled(Formsy.Form)`
+  width: 200px;
+  margin-left: auto;
+  margin-right: auto;
+`
+export default Form
